FieldReport: "In my webapp, upgrading from geo 0.31.x to 0.32.x break the build".
The problems is side-effects from changes inside of the getrandom crate.

1) Here is a dependency graph

MyApp -> MyLib -> geo -> rand -> getrandom

1) [getrandom](<https://crates.io/crates/getrandom>) The change releates to the sourcing of randomness for "supported targets"
Here is the relevant quote from the getrandoms README

 "wasm32-unknown-unknown" is no longer supported by default. see
(webassembly-support)[<https://crates.io/crates/getrandom#webassembly-support>]

For this PR the summary is

rand must enable the "wasm-js" and the root crate ( MyApp in this example )
must set this enviromment variable

RUSTFLAGS='--cfg getrandom_backend="wasm_js"'

2) [rand](https://crates.io/crates/rand>) takes its own view on this matter

Here is the relevant quote form the rand README

"The WASI and Emscripten targets are directly supported. The wasm32-unknown-unknown target is not automatically supported. To enable support for this target, refer to the getrandom documentation for WebAssembly. Alternatively, the os_rng feature may be disabled."

The best we can do, in geo, I think, is to optionally remove the rand crate, for builds that target the web.

geo = { path ="0.32", features =  [ "earcutr", "spade", "multithreading"], default-features = false }

*I should empahsise* this comes at the cost of removing the kmeans module and its associated functionality.